python/requirements.txt
curl_cffi>=0.7.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
markdown>=3.7
httpx>=0.27.0
